return bestMove instead of alpha at fail-low nodes
for some strange reason returning alpha on fail-low
destroyed the search. Instead we must return the exact score.

also change how leaf nodes are handled, by returning
a static evaluation instead of the max score of the
current move list, which won't be able to be done when
we move to a staged move generation.