add LOS to AutoPlay::printStats()
LOS (likelihood of superiority) is a simple formula to calculate the
difference in strength of two players. It's based on 50%, so if the
player that wins has an LOS of 58% then that player is 8% more likely
to win.

Currently, the autotest has the engine play against itself, so the LOS is
used to calculate an error. Theoretically, since the engine is playing
both sides using the same TT the outcome should be 50% win for both white
and black. But this is not the case with random play. Henceforth, (1-LOS)
gives the error that must be overcome to show that two different versions
of the engine are actually different in strength.