add enginetester program
enginetester is an automated engine testing program that will take as
input the number of games to run and the location of the two engines
to test. Once the games are complete summary statistics are printed
out, including win/loss/draw and the LOS for the combinations of
engine one and two playing as black and white.