Board: add debuging asserts to make/unmake
Since make/unmake directly edit the board and piece list array, making
sure that the values of the move being applied is correct is very importaint.

This adds a "#define DEBUG_MAKE_MOVE" that can be set to enable the assert
version of make/unmake which will exit the program if a move that is invalid
is trying to be applied.