add pvMove Array to store the PV
When all moves are searched because of no beta-cutoff, the node type
is eather an All-Node or a PV-Node. All-Nodes only have a score because
no move improved alpha, bue PV-Nodes did improve alpha.

By setting pvMove to null, and storing the last move that increased alpha
we can know if we're in an All or PV node by checking if pvMove is still null.