remove TT score lookup in getMovesList
scores from the transposition table should not be looked up when
generating the move list. This is because the saved scores give the
wrong values when sorting.

saved cut-nodes would have low values saved and henceforth searched last
saved all-nodes would have high values saved and henceforth searched first

This causes the search to progressively find more and more all-nodes,
which take longer to search; and stay away from cut-nodes which are faster.