rewrite TransTable
Major rewrite of the TransTable:

* Merge setScore/setBest/setPV into setItem
* Remove all debuging functions because they nolonger work with the new format
* Add getItem to retrive a single table entry
* Add getMove/getScore to TableItem

The old TransTable wasn't working the correct way.

Instead of always returning a score, it's only suppose to return a score when the
current remaining search depth is less-equal to the saved depth in the table item.
Else, we are trying to get a score for a depth we haven't searched yet.

Also, the returned score should not always be the saved score:
	if type=all-node return alpha
	if type=cut-node return beta
	if type=pv-node return min(max(score, alpha), beta)
		which is value inside alpha/beta