add rebuildHash to Board class
rebuildHash will be usefull when setting the board to a random state.
But it currently fixes a bug with the initialization of a Board instance.

The hash key of the board is set to startHash, but that value isn't
set until the hash table is created. Henceforth, a call to rebuildHash
is required after creating the hash table and before any moves are
done with the board.