add Position class and a setBoard function to the Board class
The Position class is used to parse and print board positions
and for setting a board instance from a parsed string.

The existing Position struct and all usage of it has been replaced
by the new Position class. As such, printPieceList was changed to
be a member function of Board; and getPosition now uses the new