fix compile error with gcc 4.5 and abs()
For some odd reason gcc 4.5 includes and/or uses: float abs(float)
as the abs() function instead of: int abs(int). This causes a compile
error when the value is used as an array index.

Instead, use an ABS(X) macro which bypasses gcc 4.5's use of the cmath
abs (instead of the cstdlib's version that returns int).