fix uninitialized use of MoveNode::score, MoveNode::check and TransItem::depth
gcc gives a compiler error in the getPerftMoveList because both
MoveNode::score and MoveNode::check are not initialized. Fixed by
adding a constructor to MoveNode and initializing the values to 0.

valgrind caught the uninitialized use of TransItem::depth which would
cause setItem to fail at saving when the initialized value was higher
than the value being saved.