add incheckMove and attackLine to reduce legal move checking
gprof shows that checking if the king is in check takes up much of the
engine's time. To simplify this, most of the movelist generation functions
now use incheckMove instead of incheck. This new function will only check
the line between the given move's from/to and the king. So for most cases
only 2 directions instead of 16 are checked.