make move score calculation an incremental update
Board.mscore holds the current score for the board from white's
view. Board.eval() returns the score based on the side to move
and no longer does a full recalculation of the score. Instead,
the score updating is moved to Board.make() and Board.unmake.

Board.rebuildScore() now does what Board.eval() did, but is only
called in board initialization and setting a new board position.